zustang init ok

+++ app/users-table.tsx
@@ -13,14 +13,15 @@ import { useRouter } from 'next/navigation';
 import { deleteUserById } from './dashboard/action';
 import { User } from '@/models/user/user';
 
+
 export function UsersTable({
   users,
   offset,
   sessionUser
 }: {
-  users?: User[] | undefined;
+  users?: any[];
   offset?: number | null;
-  sessionUser: User
+  sessionUser: any
 }) {
   const router = useRouter();
 
@@ -63,7 +64,7 @@ export function UsersTable({
 function UserRow({ user, sessionUser }: { user: User, sessionUser:User }) {
   const userId = user.id;
   const deleteUser = async () => {
-    await deleteUserById(userId)
+    userId && await deleteUserById(userId)
   }
 
   return (
